trading-bot-backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.121.2"}
psycopg = {extras = ["binary"], version = "^3.2.12"}
ccxt = "^4.5.19"
cryptography = "^46.0.3"
python-dotenv = "^1.2.1"
pydantic-settings = "^2.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
